Update FastText (#1)
We are seeing clang errors when compiling on OSX. Here we update fasttext package to the new API released by Facebook (see here). We eventually hope to leverage this direct dependency via pip. However, we are still seeing clang errors even with this official release. For the time-being we rely on 3rd party wheels because of clang incompatibility on OSX